update compiler-hooks to use the compiler API
