+++ tests/compiler-hooks/app/src/main/graphql/operation.graphql
