+++ tests/compiler-hooks/app/src/main/graphql/schema.graphqls
