+++ tests/compiler-hooks/app/src/main/kotlin/hooks/typenameinterface/HasTypeName.kt
