+++ tests/compiler-hooks/app/src/test/kotlin/test/CapitalizeEnumValuesTest.kt
