+++ tests/compiler-hooks/app/src/test/kotlin/test/GettersAndSettersTest.kt
