+++ tests/compiler-hooks/app/src/test/kotlin/test/TypeNameInterfaceTest.kt
