+++ tests/compiler-hooks/capitalize-enum-values/src/main/resources/META-INF/services/com.apollographql.apollo3.compiler.Plugin
@@ -1,0 +1,1 @@
+hooks.TestPlugin
