+++ tests/compiler-hooks/typename-interface/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+  id("org.jetbrains.kotlin.jvm")
+}
+
+apolloTest()
+
+dependencies {
+  implementation(libs.apollo.compiler)
+}
